# openapi.yaml
openapi: 3.0.3
info:
  title: Arrosage API
  description: Simple API for controlling the watering system
  version: 1.0.0
  contact:
    name: Arrosage System
  license:
    name: MIT
servers:
  - url: http://localhost:8000
    description: Development server
paths:
...
  /mode:
    get:
      summary: Get current mode
      description: Get the current system mode
      operationId: getCurrentMode
      tags:
        - Mode
      responses:
        '200':
          description: Current mode information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModeResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    post:
      summary: Set new mode
      description: Set a new system mode
      operationId: setCurrentMode
      tags:
        - Mode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModeRequest'
      responses:
        '200':
          description: Mode set successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModeResponse'
        '400':
          description: Invalid mode value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Request body validation error (e.g. missing or malformed fields)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /status:
    get:
      summary: Get current status
      description: Get the current sequence status
      operationId: getCurrentStatus
      tags:
        - Status
      responses:
        '200':
          description: Current status information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /pause:
    post:
      summary: Pause system
      description: Pause the watering system
      operationId: pauseSystem
      tags:
        - Control
      responses:
        '200':
          description: Pause action result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /resume:
    post:
      summary: Resume system
      description: Resume the watering system from pause
      operationId: resumeSystem
      tags:
        - Control
      responses:
        '200':
          description: Resume action result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /reset:
    post:
      summary: Reset system
      description: Reset the watering system to initial state
      operationId: resetSystem
      tags:
        - Control
      responses:
        '200':
          description: Reset action result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /start:
    post:
      summary: Start sequence
      description: Start a watering sequence (only works in SEMI_AUTO mode)
      operationId: startSystem
      tags:
        - Control
      responses:
        '200':
          description: Start action result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionResponse'
              examples:
                success:
                  summary: Successful start
                  value:
                    success: true
                    message: "Sequence started successfully"
                    current_mode: "semi_auto"
                wrongMode:
                  summary: Wrong mode
                  value:
                    success: false
                    message: "Cannot start sequence. System must be in SEMI_AUTO mode (current mode: auto)"
                    current_mode: "auto"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /relay/open:
    post:
      summary: Open a single relay
      description: Open the specified relay (only works in MANUAL mode). Any previously opened relay is closed first.
      operationId: openRelay
      tags:
        - Relay
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RelayOpenRequest'
      responses:
        '200':
          description: Relay open action result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RelayStatusResponse'
              examples:
                success:
                  summary: Relay opened successfully
                  value:
                    success: true
                    message: "Relay 2 opened successfully"
                    current_mode: "manual"
                    relays:
                      - { relay_id: 0, is_open: false }
                      - { relay_id: 1, is_open: false }
                      - { relay_id: 2, is_open: true }
                      - { relay_id: 3, is_open: false }
                      - { relay_id: 4, is_open: false }
                      - { relay_id: 5, is_open: false }
                      - { relay_id: 6, is_open: false }
                      - { relay_id: 7, is_open: false }
                wrongMode:
                  summary: Wrong mode
                  value:
                    success: false
                    message: "Cannot open relay. System must be in MANUAL mode (current mode: auto)"
                    current_mode: "auto"
                    relays:
                      - { relay_id: 0, is_open: false }
                      - { relay_id: 1, is_open: false }
                      - { relay_id: 2, is_open: false }
                      - { relay_id: 3, is_open: false }
                      - { relay_id: 4, is_open: false }
                      - { relay_id: 5, is_open: false }
                      - { relay_id: 6, is_open: false }
                      - { relay_id: 7, is_open: false }
        '400':
          description: Invalid relay_id (must be an integer between 0 and 7)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidId:
                  summary: Invalid relay_id
                  value:
                    detail: "Invalid relay_id '8'. Must be an integer between 0 and 7."
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /relay/close:
    post:
      summary: Close all relays
      description: Close all relays (only works in MANUAL mode).
      operationId: closeRelays
      tags:
        - Relay
      responses:
        '200':
          description: Relay close action result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RelayStatusResponse'
              examples:
                success:
                  summary: All relays closed
                  value:
                    success: true
                    message: "All relays closed successfully"
                    current_mode: "manual"
                    relays:
                      - { relay_id: 0, is_open: false }
                      - { relay_id: 1, is_open: false }
                      - { relay_id: 2, is_open: false }
                      - { relay_id: 3, is_open: false }
                      - { relay_id: 4, is_open: false }
                      - { relay_id: 5, is_open: false }
                      - { relay_id: 6, is_open: false }
                      - { relay_id: 7, is_open: false }
                wrongMode:
                  summary: Wrong mode
                  value:
                    success: false
                    message: "Cannot close relays. System must be in MANUAL mode (current mode: auto)"
                    current_mode: "auto"
                    relays:
                      - { relay_id: 0, is_open: false }
                      - { relay_id: 1, is_open: false }
                      - { relay_id: 2, is_open: false }
                      - { relay_id: 3, is_open: false }
                      - { relay_id: 4, is_open: false }
                      - { relay_id: 5, is_open: false }
                      - { relay_id: 6, is_open: false }
                      - { relay_id: 7, is_open: false }
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    ModeRequest:
      type: object
      required:
        - mode
      properties:
        mode:
          type: string
          enum: [manual, auto, semi_auto, pause]
          description: The mode to set
          example: "auto"
    ModeResponse:
      type: object
      required:
        - current
        - valid_modes
      properties:
        current:
          type: string
          enum: [manual, auto, semi_auto, pause]
          description: Current system mode
          example: "auto"
        valid_modes:
          type: array
          items:
            type: string
            enum: [manual, auto, semi_auto, pause]
          description: List of all valid modes
          example: ["manual", "auto", "semi_auto", "pause"]
    StatusResponse:
      type: object
      required:
        - status
        - has_active_sequence
      properties:
        status:
          oneOf:
            - type: "null"
            - $ref: '#/components/schemas/SequenceStatus'
          description: Current sequence status or null if no active sequence
        has_active_sequence:
          type: boolean
          description: Whether there is an active sequence running
          example: true
...
    ActionResponse:
      type: object
      required:
        - success
        - message
        - current_mode
      properties:
        success:
          type: boolean
          description: Whether the action was successful
          example: true
        message:
          type: string
          description: Human-readable message about the action result
          examples:
            - "System paused successfully"
            - "System resumed successfully"
            - "System reset successfully"
            - "Sequence started successfully"
        current_mode:
          type: string
          enum: [manual, auto, semi_auto, pause]
          nullable: true
          description: Current system mode after the action
          example: "pause"
    ErrorResponse:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: Error message
          example: "Invalid mode 'invalid'. Valid modes are: ['manual', 'auto', 'semi_auto', 'pause']"
    RelayOpenRequest:
      type: object
      required:
        - relay_id
      properties:
        relay_id:
          type: integer
          minimum: 0
          maximum: 7
          description: Relay identifier (0-7)
          example: 2
...